% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make9mer.R
\name{make9mer}
\alias{make9mer}
\title{make9mer}
\usage{
make9mer(protseq)
}
\arguments{
\item{protseq}{Sequence of the protein. Data.frame with: codon number in codon column, and amino acid for each codon in aminoacid column}
}
\value{
list9mers. List of length equal of protein length. Includes data frames with 9-mers.
}
\description{
9mer generator from protein sequence
}
\details{
needs a protein length at least 18 amino acid
}
\examples{
protseq <- data.frame(  codon = seq(1:20)
                      , aminoacid = c("M","E","E","P","Q","S","D","P","S","V","E","P","P","L","S","Q","E","T","F","S","D")
                      )
make9mer(protseq = protseq)
}
\author{
Nathan Lemonnier \email{nathanael.lemonnier@gmail.com}
}
