% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make9mer.R
\name{make9mer}
\alias{make9mer}
\title{make9mer}
\usage{
make9mer(protseq)
}
\arguments{
\item{protseq}{Sequence of the protein. Data.frame with: codon number in codon column, and amino acid for each codon in aminoacid column}
}
\description{
9mer generator from protein sequence of length at least 18 amino acid
}
\examples{
protseq <- data.frame(  codon = seq(1:20)
                      , aminoacid = c("A","C","D","E","F","G","H","I","K","L","M","N","P","Q","R","S","T","V","W","Y")
                      )
make9mer(protseq = protseq)
}
